util refac
